Translation-agent/Translate-agent/backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
google-generativeai
python-multipart
aiofiles
gtts
